refactor: move more types into the "types" module
