+++ src/config/models/types/mod.rs
@@ -1,11 +1,15 @@
 mod address_family;
 mod base_url;
+mod cross_origin;
 mod duration;
 mod minify;
+mod serde;
 mod ws;
 
 pub use address_family::*;
 pub use base_url::*;
+pub use cross_origin::*;
 pub use duration::*;
 pub use minify::*;
+pub use serde::*;
 pub use ws::*;
